moved plotting methods to utils/plotting, see issue #269
